refactor(client): move all component files into respective folders
