+++ client/src/components/home/home.jsx
@@ -2,16 +2,16 @@ import React, { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faBookmark, faSearch } from '@fortawesome/free-solid-svg-icons';
-import { selectSavedImages, selectSearchedImages } from '../slices/savedImagesSelectedSlice';
-import MusicPlayer from './musicPlayer';
-import LoginModal from './login/loginModal';
-import SongContainer from './songs/songContainer';
-import ImageContainer from './images/imageContainer';
-import Search from './search';
-import useAuth from '../useAuth';
-import SpotifyLoginButton from './spotifyLogin';
-import LoginIcon from './icons_and_holders/loginIcon';
-import LogoutIcon from './icons_and_holders/logoutIcon';
+import { selectSavedImages, selectSearchedImages } from '../../slices/savedImagesSelectedSlice';
+import MusicPlayer from '../musicPlayer/musicPlayer';
+import LoginModal from '../login/loginModal';
+import SongContainer from '../songs/songContainer';
+import ImageContainer from '../images/imageContainer';
+import Search from '../search/search';
+import useAuth from '../../useAuth';
+import SpotifyLoginButton from '../spotifyLogin/spotifyLogin';
+import LoginIcon from '../icons_and_holders/loginIcon';
+import LogoutIcon from '../icons_and_holders/logoutIcon';
 import styles from './home.module.css';
 
 const code = new URLSearchParams(window.location.search).get('code');
